[DESK-51] Moving the workflow configs
