--- src/models/settings/index.js
@@ -1,7 +1,0 @@
-const Collection = require('./Collection');
-const CollectionField = require('./CollectionField');
-
-module.exports = {
-    Collection,
-    CollectionField
-};
